Rename module to resolve naming conflict
