+++ src/email_draft_generator/__main__.py
@@ -1,0 +1,4 @@
+import email_draft_generator.main
+
+if __name__ == "__main__":
+	email_draft_generator.main.main()
