+++ src/email_draft_generator/file_parser/__main__.py
@@ -1,0 +1,4 @@
+import email_draft_generator.file_parser.main
+
+if __name__ == "__main__":
+	email_draft_generator.file_parser.main.main()
