+++ src/email_draft_generator/file_parser/main.py
