+++ src/email_draft_generator/gmail.py
