+++ src/email_draft_generator/mail_util.py
