+++ src/email_draft_generator/main.py
@@ -2,9 +2,9 @@ import json
 import concurrent.futures
 import fileinput
 
-import email_generator.gmail
-import email_generator.mail_util
-import email_generator.email_creator
+import email_draft_generator.gmail
+import email_draft_generator.mail_util
+import email_draft_generator.email_creator
 
 def main():
 	# File paths
@@ -23,16 +23,16 @@ def main():
 
 	# Authenticate with Google
 	print("Authenticating")
-	creds = email_generator.gmail.get_creds(google_token_path, google_oauth_credentials_path)
+	creds = email_draft_generator.gmail.get_creds(google_token_path, google_oauth_credentials_path)
 
 	# Build all of the attachments
 	print("Processing attachments")
 	attachments = []
 	with concurrent.futures.ProcessPoolExecutor() as executor:
 		for attachment_path in attachment_paths:
-			attachments.append(email_generator.mail_util.build_attachment(attachment_path))
+			attachments.append(email_draft_generator.mail_util.build_attachment(attachment_path))
 
 	print("Generating and uploading E-mails")
 	with concurrent.futures.ProcessPoolExecutor() as executor:
 		for company in companies:
-			email_generator.gmail.create_draft(creds, email_generator.email_creator.create_email_body(company, attachments))
+			email_draft_generator.gmail.create_draft(creds, email_draft_generator.email_creator.create_email_body(company, attachments))
