--- src/email_generator/__main__.py
@@ -1,4 +1,0 @@
-import email_generator.main
-
-if __name__ == "__main__":
-	email_generator.main.main()
