--- src/email_generator/file_parser/__main__.py
@@ -1,4 +1,0 @@
-import email_generator.file_parser.main
-
-if __name__ == "__main__":
-	email_generator.file_parser.main.main()
